List of included changes:   - Update ML Kit GenAI sample app to use renamed schema-compiler SDK.
PiperOrigin-RevId: 934110695
