+++ android/genai/app/build.gradle
@@ -63,7 +63,7 @@ dependencies {
   implementation 'com.google.mlkit:genai-proofreading:1.0.0-beta2'
   implementation 'com.google.mlkit:genai-rewriting:1.0.0-beta2'
   implementation 'com.google.mlkit:genai-summarization:1.0.0-beta2'
-  ksp 'com.google.mlkit:genai-structured-compiler:1.0.0-alpha1'
+  ksp 'com.google.mlkit:genai-schema-compiler:1.0.0-alpha1'
 
   implementation 'androidx.concurrent:concurrent-futures:1.3.0'
   implementation 'androidx.preference:preference-ktx:1.2.1'
@@ -79,4 +79,5 @@ dependencies {
   androidTestImplementation 'androidx.test.ext:junit:1.2.1'
   androidTestImplementation 'com.google.truth:truth:1.1.3'
   androidTestImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.8.1'
+  androidTestImplementation 'androidx.lifecycle:lifecycle-viewmodel-savedstate:2.6.2'
 }
